requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
bcrypt==5.0.0
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
Django==6.0.4
djangorestframework==3.17.1
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.135.3
greenlet==3.4.0
h11==0.16.0
idna==3.11
mysqlclient==2.2.8
openpyxl==3.1.5
passlib==1.7.4
pillow==12.2.0
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.24
reportlab==4.4.10
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
